[apps] turn size_t into int

+++ apps/shared/record_name_helper.h
@@ -82,8 +82,8 @@ private:
   typedef struct {
     const char * const * namePrefixes;
     const char * extension;
-    size_t prefixRepetitions;
-    size_t numberOfElements;
+    int prefixRepetitions;
+    int numberOfElements;
   } ReservedExtension;
 
   constexpr static ReservedExtension k_reservedExtensions[] = {
@@ -91,7 +91,7 @@ private:
     {Regression::Store::k_columnNames, Ion::Storage::lisExtension, DoublePairStore::k_numberOfSeries, sizeof(Regression::Store::k_columnNames) / sizeof(char *)},
     {Statistics::Store::k_columnNames, Ion::Storage::lisExtension, DoublePairStore::k_numberOfSeries, sizeof(Statistics::Store::k_columnNames) / sizeof(char *)}
   };
-  constexpr static size_t k_reservedExtensionsLength = sizeof(k_reservedExtensions) / sizeof(ReservedExtension);
+  constexpr static int k_reservedExtensionsLength = sizeof(k_reservedExtensions) / sizeof(ReservedExtension);
 
   /* Returns true if the name is reserved for an other extension than the one
   * given in parameter. Return false if the name is reserved for this extension
